requirements.txt
autogen==0.2.2
pytest==8.0.0
ruff==0.2.1
mypy==1.8.0
python-dotenv==1.0.1 